refactor inventory to use shared method after credential resolution
